Fix loadWheel showing
Eliminate the delay of the show() command for the loadWheel image

Ensure the image is shown when reset is called by adding a 100 ms delay
to the click function before proceeding to the resetMenu function.

Ensure that the reset only tries to occur when some JSON has already
been loaded.